+++ messaging/build.gradle
@@ -17,6 +17,7 @@ dependencies {
 
     implementation 'org.springframework.boot:spring-boot-starter'
     api 'org.springframework.kafka:spring-kafka'
+    api 'org.hibernate:hibernate-spatial:6.6.0.Final'
 
     compileOnly 'org.projectlombok:lombok'
     annotationProcessor 'org.projectlombok:lombok'
